requirements.txt
argon2-cffi==18.1.0
authy==2.2.1
Babel==2.6.0
bcrypt==3.1.4
boto==2.49.0
boto3==1.7.84
botocore==1.10.84
cachetools==2.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
defusedxml==0.5.0
Django==2.1
django-allauth==0.36.0
django-bootstrap3==10.0.1
django-phonenumber-field==2.0.1
django-rest-auth==0.9.3
django-storages==1.6.6
djangorestframework==3.8.2
docutils==0.14
Faker==0.8.15
futures==3.1.1
google-api-core==1.3.0
google-assistant-grpc==0.2.0
google-assistant-sdk==0.5.0
google-auth==1.5.0
google-auth-oauthlib==0.2.0
google-cloud-core==0.28.1
google-cloud-storage==1.10.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.13.0
gunicorn==19.9.0
idna==2.7
jmespath==0.9.3
mysqlclient==1.3.13
oauthlib==2.1.0
pathlib2==2.3.2
phonenumbers==8.8.4
phonenumberslite==8.9.10
Pillow==5.1.0
protobuf==3.6.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycparser==2.18
PyJWT==1.6.4
PySocks==1.6.8
python-dateutil==2.7.3
python-dotenv==0.7.1
python3-openid==3.1.0
pytz==2018.4
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
s3transfer==0.1.13
sendotp==0.0.4
simplejson==3.16.0
six==1.11.0
sounddevice==0.3.11
tenacity==4.12.0
text-unidecode==1.2
twilio==6.16.2
urllib3==1.23
